+++ pydidas/data_io/implementations/hdf5_io.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@ Module with the Hdf5Io class for importing and exporting Hdf5 data.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -44,7 +44,7 @@ from .io_base import IoBase
 
 
 class Hdf5Io(IoBase):
-    """IObase implementation for Hdf5 files."""
+    """IoBase implementation for Hdf5 files."""
 
     extensions_export = HDF5_EXTENSIONS
     extensions_import = HDF5_EXTENSIONS
@@ -60,44 +60,55 @@ class Hdf5Io(IoBase):
         ----------
         filename : Union[pathlib.Path, str]
             The filename of the file with the data to be imported.
-        dataset : str, optional
-            The full path to the hdf dataset within the file. The default is
-            "entry/data/data".
-        slicing_axes : list, optional
-            The axes to be slices by the specified frame indices. The default
-            is [0].
-        frame : Union[int, list], optional
-            The indices of the unused axes to identify the selected dataset.
-            Integer values will be interpreted as values for axis 0.
-            The default is 0.
-        roi : Union[tuple, None], optional
-            A region of interest for cropping. Acceptable are both 4-tuples
-            of integers in the format (y_low, y_high, x_low, x_high) as well
-            as 2-tuples of integers or slice objects. If None, the full image
-            will be returned. The default is None.
-        returnType : Union[datatype, 'auto'], optional
-            If 'auto', the image will be returned in its native data type.
-            If a specific datatype has been selected, the image is converted
-            to this type. The default is 'auto'.
-        binning : int, optional
-            The rebinning factor to be applied to the image. The default
-            is 1.
+        **kwargs : dict
+            The following kwargs are supported:
+
+            dataset : str, optional
+                The full path to the hdf dataset within the file. The default is
+                "entry/data/data".
+            slicing_axes : list, optional
+                The axes to be slices by the specified frame indices. The default
+                is [0].
+            frame : Union[int, list], optional
+                The indices of the unused axes to identify the selected dataset.
+                Integer values will be interpreted as values for axis 0.
+                The default is 0.
+            roi : Union[tuple, None], optional
+                A region of interest for cropping. Acceptable are both 4-tuples
+                of integers in the format (y_low, y_high, x_low, x_high) or
+                2-tuples of integers or slice objects. If None, the full image
+                will be returned. The default is None.
+            returnType : Union[datatype, 'auto'], optional
+                If 'auto', the image will be returned in its native data type.
+                If a specific datatype has been selected, the image is converted
+                to this type. The default is 'auto'.
+            binning : int, optional
+                The re-binning factor to be applied to the image. The default
+                is 1.
+            auto_squeeze : bool, optional
+                Flag to automatically squeeze data dimensions before returning
+                the data. The default is True.
 
         Returns
         -------
         data : pydidas.core.Dataset
             The data in form of a pydidas Dataset (with embedded metadata)
         """
-        frame = kwargs.get("frame", 0)
+        frame = kwargs.get("frame", [])
         if isinstance(frame, Integral):
             frame = [frame]
         dataset = kwargs.get("dataset", "entry/data/data")
-        slicing_axes = kwargs.get("slicing_axes", [0])
+        slicing_axes = kwargs.get("slicing_axes", [])
+        if isinstance(slicing_axes, Integral):
+            slicing_axes = [slicing_axes]
 
-        if len(frame) < len(slicing_axes):
+        auto_squeeze = kwargs.get("auto_squeeze", True)
+
+        if len(frame) != len(slicing_axes):
             raise ValueError(
-                "The number of frames must not be shorter than "
-                "the number of slicing indices."
+                "The length of frame items must be equal to the number of "
+                f"slicing indices. Given frames: `{frame}`; given slicing axes: "
+                f"`{slicing_axes}`."
             )
 
         if len(slicing_axes) == 0:
@@ -109,7 +120,9 @@ class Hdf5Io(IoBase):
                 for _i in range(amax(slicing_axes) + 1)
             ]
         with CatchFileErrors(filename):
-            _data = squeeze(read_hdf5_dataset(filename, dataset, _slicer))
+            _data = read_hdf5_dataset(filename, dataset, _slicer)
+            if auto_squeeze:
+                _data = squeeze(_data)
         cls._data = Dataset(
             _data,
             metadata={"slicing_axes": slicing_axes, "frame": frame, "dataset": dataset},
@@ -119,7 +132,7 @@ class Hdf5Io(IoBase):
     @classmethod
     def export_to_file(cls, filename: Union[Path, str], data: ndarray, **kwargs: dict):
         """
-        Export data to an Hdf5 file.
+        Export data to a Hdf5 file.
 
         Parameters
         ----------
@@ -127,12 +140,14 @@ class Hdf5Io(IoBase):
             The filename
         data : np.ndarray
             The data to be written to file.
-        dataset : str, optional
-            The full path to the hdf dataset within the file. The default is
-            "entry/data/data".
-        overwrite : bool, optional
-            Flag to allow overwriting of existing files. The default is False.
+        **kwargs : dict
+            Additional keyword arguments. Supported keyword arguments are:
 
+            dataset : str, optional
+                The full path to the hdf dataset within the file. The default is
+                "entry/data/data".
+            overwrite : bool, optional
+                Flag to allow overwriting of existing files. The default is False.
         """
         cls.check_for_existing_file(filename, **kwargs)
         dataset = kwargs.get("dataset", "entry/data/data")