+++ pydidas_plugins/input_plugins/hdf5_file_series_loader.py
@@ -44,7 +44,7 @@ class Hdf5fileSeriesLoader(InputPlugin):
     start index.
 
     The final filename is
-    <SCAN base directory>/<SCAN name pattern with index subsituted for hashes>.
+    <SCAN base directory>/<SCAN name pattern with index substituted for hashes>.
 
     The dataset in the Hdf5 file is defined by the hdf5_key Parameter.
 
@@ -67,11 +67,16 @@ class Hdf5fileSeriesLoader(InputPlugin):
     basic_plugin = False
     plugin_type = INPUT_PLUGIN
     default_params = get_generic_param_collection(
-        "hdf5_key", "images_per_file", "file_stepping", "_counted_images_per_file"
+        "hdf5_key",
+        "hdf5_slicing_axis",
+        "images_per_file",
+        "file_stepping",
+        "_counted_images_per_file",
     )
     input_data_dim = None
     output_data_dim = 2
     advanced_parameters = InputPlugin.advanced_parameters.copy() + [
+        "hdf5_slicing_axis",
         "images_per_file",
         "file_stepping",
     ]
@@ -83,12 +88,10 @@ class Hdf5fileSeriesLoader(InputPlugin):
         InputPlugin.pre_execute(self)
         _i_per_file = self.get_param_value("images_per_file")
         if _i_per_file == -1:
-            _n_per_file = get_hdf5_metadata(
+            _i_per_file = get_hdf5_metadata(
                 self.get_filename(0), "shape", dset=self.get_param_value("hdf5_key")
-            )[0]
-            self.set_param_value("_counted_images_per_file", _n_per_file)
-        else:
-            self.set_param_value("_counted_images_per_file", _i_per_file)
+            )[self.get_param_value("hdf5_slicing_axis")]
+        self.set_param_value("_counted_images_per_file", _i_per_file)
 
     def get_frame(self, frame_index: int, **kwargs: dict) -> tuple[Dataset, dict]:
         """
@@ -114,6 +117,7 @@ class Hdf5fileSeriesLoader(InputPlugin):
         kwargs["dataset"] = self.get_param_value("hdf5_key")
         kwargs["frame"] = _hdf_index
         kwargs["binning"] = self.get_param_value("binning")
+        kwargs["slicing_axes"] = [self.get_param_value("hdf5_slicing_axis")]
         _data = import_data(_fname, **kwargs)
         _data.axis_units = ["pixel", "pixel"]
         _data.axis_labels = ["detector y", "detector x"]