+++ tests/apps/test_composite_creator_app.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -70,7 +70,7 @@ class TestCompositeCreatorApp(unittest.TestCase):
         cls._q_settings = PydidasQsettings()
         cls._border = cls._q_settings.value("user/mosaic_border_width", int)
         cls._bgvalue = cls._q_settings.value("user/mosaic_border_value", float)
-        _mask = np.zeros((cls._img_shape), dtype=np.bool_)
+        _mask = np.zeros(cls._img_shape, dtype=np.bool_)
         _maskfile = cls._path.joinpath("mask.npy")
         np.save(_maskfile, _mask)
         cls._maskfile = _maskfile
@@ -280,7 +280,9 @@ class TestCompositeCreatorApp(unittest.TestCase):
     def test_multiprocessing_func(self):
         app = self.get_default_app()
         app._config["current_fname"] = self._hdf5_fnames[0]
-        app._config["current_kwargs"] = dict(dataset="/entry/data/data", frame=0)
+        app._config["current_kwargs"] = dict(
+            dataset="/entry/data/data", frame=0, slicing_axes=[0]
+        )
         _image = app.multiprocessing_func(0)
         self.assertTrue((_image == self._data[0]).all())